feat: allow youth card renewal in pending state (#197)
Previously guardians have inadvertently tried to register their youths
multiple times. If the youth card was registered succesfully but not yet
accepted, any subsequent registration attempts have been met with a
non-descriptive error message, which has resulted in confusion and calls
to the Nutakortti contact person and youth clubs.

After this change guardians can register the same youth multiple times
as long as the youth's status is either "expired / tunnus vanhentunut"
or "pending / kotisoitto tekemättä". This will just overwrite the
existing youth with the provided registration form data. It is still not
possible to overwrite data for youths with any other status.